+++ packages/updating-element/CHANGELOG.md
@@ -22,3 +22,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - UpdatingElement moved from `lit-element` package to `updating-element` package.
+- Console warnings added for removed API and other element problems in developer mode. Set `static strictWarnings` to receive warnings for common mistakes; for example, when an update triggers another update.